converted long to BigInteger

+++ src/main/java/BezoutResult.java
@@ -1,1 +1,2 @@
-public record BezoutResult(long gcd, long x, long y) {}
+import java.math.BigInteger;
+public record BezoutResult(BigInteger gcd, BigInteger x, BigInteger y) {}
